+++ src/collective/contact/core/browser/contact.py
@@ -51,6 +51,3 @@ class Contact(BaseView):
 
         organization = held_position.get_organization()
         self.organizations = organization and organization.get_organizations_chain() or []
-
-        # also show fields that were added TTW
-        self.ttw_fields = get_ttw_fields(held_position)